Adjust API and Add tests
Lots done here as I keep rejiggering the api and feel to better match
the other agents. I'm not going to run down the whole thing in the
commit message because it is still in flux. Most notably, I have spans
as responsible for their own tags, rather than calling & relying on the
request to organize it.