Loosen version requirement for ralouphie/getallheaders
Some users have 3.x required in their Laravel apps, and both 2.x and 3.x work fine. There's little to no chance of backwards incompatible changes in this library because it's meant to polyfill a single, defined-elsewhere function. Even the 2.x to 3.x was just dropping support for very old PHPs.