Updates CoreAgent to 1.2.4
A few bug fixes, but most notable fix is better parsing/naming of SQL
emitted in real-life by Laravel. So instead of having "SQL" as the name
of the row in the web UI, it will say "User/Select" or "Post/Update" and
so on more often.

Note: it wasn't skipping any instrumented SQL previously, just was
falling back to a generic name for it.