Log invalid log levels as errors and default to 'debug'
Prior to this change, an invalid log level passed to the FilteredLogLevelDecorator
constructor would result in an InvalidArgumentException. This means that a mis-
configured production system could fail to boot.

This patch proposes that an invalid log level should not be a hard failure. The
change logs the invalid log level to the 'realLogger' as an error, then sets the
default log level to 'debug'.